requirements.txt
requests
python-dateutil
streamlit
plotly